refactor: unify duplicate JWT verification middlewares
- Merged authMiddleware.js and verifyToken.js into single file
- Standardized req.user across all routes
- Updated imports in all route files
- Removed duplicate verifyToken.js file

Fixes #346
