+++ backend/middlewares/authMiddleware.js
@@ -1,47 +1,50 @@
 import jwt from "jsonwebtoken";
 import User from "../models/user.js";
 
+// Unified JWT middleware (merges protect + verifyToken)
 export const protect = async (req, res, next) => {
   try {
-    const authHeader = req.headers.authorization;
+      const authHeader = req.headers.authorization;
 
-    // Validate header format
-    if (!authHeader || !authHeader.startsWith("Bearer ")) {
-      return res.status(401).json({
-        message: "Authorization token missing or malformed",
-      });
-    }
+          if (!authHeader || !authHeader.startsWith("Bearer ")) {
+                return res.status(401).json({
+                        message: "Authorization token missing or malformed",
+                              });
+                                  }
 
-    const token = authHeader.split(" ")[1];
+                                      const token = authHeader.split(" ")[1];
 
-    // Verify JWT secret exists
-    if (!process.env.JWT_SECRET) {
-      return res.status(500).json({
-        message: "JWT secret is not configured",
-      });
-    }
+                                          if (!process.env.JWT_SECRET) {
+                                                return res.status(500).json({
+                                                        message: "JWT secret is not configured",
+                                                              });
+                                                                  }
 
-    // Verify token
-    const decoded = jwt.verify(token, process.env.JWT_SECRET);
+                                                                      const decoded = jwt.verify(token, process.env.JWT_SECRET);
 
-    // Fetch latest user data
-    const user = await User.findById(decoded.id).select("-password");
+                                                                          const user = await User.findById(decoded.id).select("-password");
 
-    // Ensure user still exists
-    if (!user) {
-      return res.status(401).json({
-        message: "User no longer exists",
-      });
-    }
+                                                                              if (!user) {
+                                                                                    return res.status(401).json({
+                                                                                            message: "User not found",
+                                                                                                  });
+                                                                                                      }
 
-    req.user = user;
+                                                                                                          // Standardized req.user for ALL routes
+                                                                                                              req.user = {
+                                                                                                                    _id: user._id,
+                                                                                                                          id: user._id,
+                                                                                                                                role: user.role,
+                                                                                                                                    };
 
-    next();
-  } catch (err) {
-    console.error("Auth middleware error:", err.message);
+                                                                                                                                        next();
+                                                                                                                                          } catch (err) {
+                                                                                                                                              console.error("Auth middleware error:", err.message);
+                                                                                                                                                  return res.status(401).json({
+                                                                                                                                                        message: "Invalid or expired token",
+                                                                                                                                                            });
+                                                                                                                                                              }
+                                                                                                                                                              };
 
-    return res.status(401).json({
-      message: "Invalid or expired token",
-    });
-  }
-};
+                                                                                                                                                              // Alias so verifyToken imports still work
+                                                                                                                                                              export const verifyToken = protect;
